Game과 Quiz의 메시지 브로커 경로 분리
* 자바의 Solid원칙의 단일 책임에 따라서 Game은 Game관련 요청만 수행. Quiz는 Quiz관련 요청만 수행을 위한 클래스 분리 진행함
